End_project/requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
Flask_CKEditor==0.4.6
Flask_Login==0.6.3
flask_sqlalchemy==3.0.5
Flask_WTF==1.2.1
WTForms==3.0.1
SQLAlchemy==2.0.19
Markdown~=3.5.1
bleach~=6.1.0
Flask-Limiter==3.5.0
pycryptodome==3.20.0
